feat:添加了redis缓存文章详情，热门文章，浏览量定时更新的功能

+++ src/main/java/com/cn/blogsystem/entity/Article.java
@@ -13,7 +13,7 @@ public class Article {
     private String title;//文章标题
     private String content;//文章内容
     private String summary;//文章摘要
-    private Integer viewCount;//浏览次数
+    private Long viewCount;//浏览次数
     @TableField(fill = FieldFill.INSERT)
     private LocalDateTime createTime;
     private Long userId;//用户id
