+++ src/main/java/com/cn/blogsystem/vo/ArticleDetailVO.java
@@ -9,7 +9,7 @@ public class ArticleDetailVO {
     private String title;//文章标题
     private String content;//文章内容
     private String summary;//文章摘要
-    private Integer viewCount;//浏览次数
+    private Long viewCount;//浏览次数
     private String createTime;
     private Long userId;//用户id
 }
